separate bridge review logic from bridge logic
