+++ src/app/bridge/_store/index.ts
@@ -1,3 +1,4 @@
+import { TxHash } from '@/blockchain_api/functions/icp/bridge_transactions';
 import { BridgeOption } from '@/blockchain_api/functions/icp/get_bridge_options';
 import { EvmToken, IcpToken } from '@/blockchain_api/types/tokens';
 import { PendingTransaction } from '@/common/helpers/session';
@@ -32,14 +33,14 @@ interface BridgeState {
   txStep: TxStepType;
   txErrorMessage: string | undefined;
   pendingTx: PendingTransaction | undefined;
+  txHash: TxHash | undefined;
+  withdrawalId: string | undefined;
 }
 
 type Action = {
   actions: {
     setActiveStep: (step: number) => void;
     setSelectedTokenType: (type: SelectionType) => void;
-    setFromToken: (token: TokenType | undefined) => void;
-    setToToken: (token: TokenType | undefined) => void;
     setAmount: (amount: string) => void;
     setSelectedOption: (option: BridgeOption) => void;
     setBridgePairs: (bridgePairs: (EvmToken | IcpToken)[]) => void;
@@ -48,9 +49,15 @@ type Action = {
     setToWalletAddress: (walletAddress: string) => void;
     setToWalletValidationError: (toWalletValidationError: string) => void;
     setSelectedTokenBalance: (tokenBalance: string) => void;
+    // select token actions
+    setFromToken: (token: TokenType | undefined) => void;
+    setToToken: (token: TokenType | undefined) => void;
+    // tx actions
     setTxStep: (step: TxStepType) => void;
     setTxErrorMessage: (err: string | undefined) => void;
     setPendingTx: (pendingTxs: PendingTransaction | undefined) => void;
+    setTxHash: (txHash: TxHash | undefined) => void;
+    setWithdrawalId: (withdrawalId: string | undefined) => void;
   };
 };
 
@@ -76,6 +83,8 @@ export const useBridgeStore = create<BridgeState & Action>()((set) => ({
   selectedTokenBalance: '',
   txErrorMessage: undefined,
   pendingTx: undefined,
+  txHash: undefined,
+  withdrawalId: undefined,
   actions: {
     setActiveStep: (activeStep) => set({ activeStep }),
     setSelectedTokenType: (selectedTokenType) => set({ selectedTokenType }),
@@ -89,8 +98,11 @@ export const useBridgeStore = create<BridgeState & Action>()((set) => ({
     setToWalletAddress: (toWalletAddress) => set({ toWalletAddress }),
     setToWalletValidationError: (toWalletValidationError) => set({ toWalletValidationError }),
     setSelectedTokenBalance: (selectedTokenBalance) => set({ selectedTokenBalance }),
+    // tx actions
     setTxStep: (txStep) => set({ txStep }),
     setTxErrorMessage: (txErrorMessage) => set({ txErrorMessage }),
+    setTxHash: (txHash) => set({ txHash }),
+    setWithdrawalId: (withdrawalId) => set({ withdrawalId }),
     setPendingTx: (pendingTx) => set({ pendingTx }),
   },
 }));
